Упразднил фильтры для сервисов

+++ Backend/ForumPOF/Application/Services/CategoriesService.cs
@@ -26,8 +26,8 @@ public class CategoriesService(
 
     public async Task<Result<Ulid>> CreateCategory(CategoryCreateRequest categoryRequest)
     {
-        //if (await _categoryRepository.CategoryExistByName(categoryRequest.Name))
-        //    return Result<Ulid>.BadRequest("Категория уже создана");
+        if (await _categoryRepository.CategoryExistByName(categoryRequest.Name))
+            return Result<Ulid>.BadRequest("Категория уже создана");
 
         var category = Category.Create(Ulid.NewUlid(), categoryRequest.Name, DateTime.Now);
 
@@ -40,11 +40,11 @@ public class CategoriesService(
     
     public async Task<Result> UpdateCategory(Ulid categoryId, CategoryUpdateRequest categoryRequest)
     {
-        //if (!await _categoryRepository.CategoryExistById(categoryId))
-        //    return Result.NotFound("Категория не найдена"); 
+        if (!await _categoryRepository.CategoryExistById(categoryId))
+            return Result.NotFound("Категория не найдена");
 
-        //if (await _categoryRepository.CategoryExistByName(categoryRequest.Name))
-        //    return Result.BadRequest("Категория уже создана");
+        if (await _categoryRepository.CategoryExistByName(categoryRequest.Name))
+            return Result.BadRequest("Категория уже создана");
 
         var category = await _categoryRepository.GetCategoryById(categoryId);
 
@@ -59,10 +59,9 @@ public class CategoriesService(
 
     public async Task<Result> DeleteCategory(Ulid categoryId)
     {
-        //if (!await _categoryRepository.CategoryExistByName(categoryName))
-        //    return Result.NotFound("Категория не найдена");
+        if (!await _categoryRepository.CategoryExistById(categoryId))
+            return Result.NotFound("Категория не найдена");
 
-        //var category = await _categoryRepository.GetCategoryByName(categoryName);
         var category = await _categoryRepository.GetCategoryById(categoryId);
 
         var isUpdated = await _categoryRepository.DeleteCategory(category);
